Emulator - Path Map
- se cambio de long a double las coordenadas del location
- refactor transportLocationUpdateDTO a SessionUpdateDTO
- se agrego el id a transportDTO
- se creo la clase TransportListDTO para devolver en vez de transportDTO[]
- nuevo converter de Transport a TransportDTO
- nuevo metodo en el servicio : findSession(id) , usado por getPath()
- termino+mejoras en DefaultEmulator
- refactor de nombres de paquetes en Web.transports
- Nuevo controller para las llamadas JSON de Transports
- nuevo jsp path que muestra el recorrido que llevò el session indicado por parametro

+++ masterProject/api/src/main/java/com/uade/pfi/api/dto/LocationDTO.java
@@ -8,27 +8,27 @@ package com.uade.pfi.api.dto;
  *
  */
 public class LocationDTO {
-	private Float latitude;
-	private Float longitude;
+	private double latitude;
+	private double longitude;
 
 	public LocationDTO() {
 	}
 	
-	public LocationDTO(float latitude, float longitude){
+	public LocationDTO(double latitude, double longitude){
 		this.latitude = latitude;
 		this.longitude = longitude;
 	}
 	
-	public void setLatitude(Float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
-	public void setLongitude(Float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
-	public Float getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
-	public Float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
